move the Term bilder to its own file
